Update send logic of DisruptorCommandBus
Update send logic of DisruptorCommandBus to set up a
CommandHandlingEntry with a Callable. The callable in this case is
actual repository.load().handle() call which regular AbstractRepository
implementations perform. This is necessary to be able to pass this send
action along to another thread.
To this end, a resetAsCallable() function has been added to the
CommandHandlingEntry.

#220