Improved matching of generics in AnnotationHandlerBeanPostProcessors
When a class would declare both `@CommandHandler` and `@QueryHandler`
annotated method, it would be proxied twice with only generics as the
main difference. This improvement ensures that the proxies method will
call the correct adapter, taking generics into account as well. It does
so by attempting to find a non-synthetic and non-abstract method with
parameters matching the invocation. If such method is found, the
corresponding adapter is invoked.

Resolves #670