Introduce a Serializer to the QuartzEventScheduler
Introduce a Serializer to the QuartzEventScheduler, as currently java
serialization is used to de-/serialize scheduled events. This is
somewhat risky and should be configurable for users.
As such, add a Serializer field defaulting to Xstream, a setter for the
serializer and a static key to store the serializer as job context.
Additionally, remove any warnings and perform a reindent of the file.

#749