Fix issues CROZ was experiencing by introducing a ExecutorService
An issues arose if a thread would perform a 'send and wait' to dispatch
a command, which would end up another node which would ping back a
'send and wait' to the original dispatcher. By introducing an
ExecutorService which instantiates a thread to receive a message, we
unblock this problem. Added a test to verify and solve the pointed out
issue.

#777