Removed the @EnableAxon annotation
This annotation was introduced before Spring Boot Autoconfiguration
support. It mimics some of the Autoconfiguration behavior, but doesn't
require Spring Boot. As this style of configuration has a lot of
(negative) side effects related to timing of initialization, we need to
discourage using this approach. First step is to remove the annotation.

Issue #738