Refactoring of CommandHandlers in MetaModel
Instead of addressing CommandHandlers using a mapping form command name
to handler, it is now a list, making it possible for multiple Command
Handlers to be defined for the same command name within an Aggregate.

This API change makes it possible to implement aggregates where
different handlers could be invoked, depending on the state of the
aggregate, taking polymorphism into account.