Added connectivity to AxonServer as default behavior of Config API
When using the Configuration API, and the AxonServer Connector is on the
classpath, it will be used by default, using sensible (dev) settings.

When starting with default settings, the AxonServer Connector will
output a message stating where AxonServer can be downloaded.

This commit also includes a few minor naming changes on the AxonServer
configuration.