Allow registration of sagafixture start recording callbacks
This change enables SagaFixture to register callbacks to be notified when the
recording of fixture execution results starts.

This is particularly helpful to testing scenarios where a Saga collaborator
resource is injected which is not a well known Axon collaborator (such as e.g.
CommandGateway), but is interested in having its interactions verified, local
to a specific Saga message handling.