Override readEventData(String, long)
Override the readEventData(String, long) so that we can set the
'fetchUntilEmpty' to true. This resolves the issue that a
Jdbc database has different ways of limiting the result when querying a
batch. As the default readEventData(String, long) implementation does
not proceed with fetching more after the initial set was empty, that
operation ignored potential mismatch in the number of events retrieved
when performing a batch query through JDBC if there were gaps present in
 the sequence numbers. This could thus cause the EventStreamSpliterator
 to think it was done with retrieving the expected amount of events for
 the given batch, whilst in fact some events were omitted.

#911