Cherry pick resolution on gap detection in JDBC from 4.0.x to 3.4.x
Cherry pick the solution to detect if there is a large gap in retrieving
 a batch of events if the global index gap is to large in JDBC.
Had to do some work to copy it over, as (1) the class has been moved and
 (2) the constructors were removed in Axon 4.
Original message:

Added extra query to detect large gaps in globalSequence

This extra query helps streams to "jump" gaps larger than the batchSize.
In environments where these gaps inexistent, it is possible to disable
it. This could be in cases where batch sizes are executed through a
TOP/LIMIT clause.

Resolves #419