Cherry pick scheduled message timestamps issue from 4.0.x to 3.4.x
Cherry pick the solution which ensures that the timestamp which is used
for scheduled events/deadlines is as expected when verifying the output.
Additionally, I had to copy over the adjustments to the
ScheduledDeadlineInfo class to store a DeadlineMessage i.o. an object,
as the latter approach would update the timestamp. This change was made
to accompany the registration of handler/dispatch interceptors for
deadline messages.
Original message:

Scheduled events should use trigger timestamp

In the test fixtures, scheduled events (and deadlines) should use the
simulated trigger time as the timestamp of event messages, instead of
the current time. This commit fixes the timestamp of the
StubEventScheduler, StubDeadlineManager as well as the
SimpleDeadlineManager.

Resolves #882