Changed in-memory event storage engine to reflect changes to opened streams
Most collections in the collections API provide a weakly consistent view
which does not update elements in an iterator when new items are added.
This was also the case in the InMemoryEventStore (which is not intended
for production use, anyway).

This change ensures that changes are properly reflected to streams that
have already been opened.