Adjust command and query processing
Adjust the way commands and queries are processed by the
AxonServerCommandBus and AxonServerQueryBus. Instead of using a queue
which is periodically polled for new commands/queries, instead use a
ThreadPoolExecutor which receives a private Runnable implementation to
process a command/query.
Additionally, upgrade the ProcessingInstructionHelper to an actual
utility class, with javadoc attached to it.