Added explicit support for Optional in Queries
On the handling side, Optional is considered a wrapper of the actual
return type. Hence, a handler declaring Optional<T> as return type is
considered a handler returning T.

On the dispatching side, Optional is merely used to wrap (potentially
null) return values. It doesn't change the expectations of the query
handler at all.

Resolves #552