Return an empty MetaDataValue of null
If the GrpcMetaDataConverter#convertToMetaDataValue(Object) methods
receives null, the serializer will change the type to 'empty'. When such
 a MetaDataValue is pulled through the
 convertFromMetaDataValue (MetaDataValue), the Serializer will fail.
Thus, it would be beneficial if the convertTo method does not set the
null serialized object, but just return an empty MetaDataValue instance.
Additionally, the convertTo and convertFrom functionality should be
tested.