Make the ScheduledExecutorService configurable
Currently, the AxonServerConnectionManager hard codes the
ScheduledExecutorService being used, whilst there are several use cases
imaginable where the users needs to specify their own executor services.
To be able to configure this easily, with a set default, introduce the
Builder pattern for the AxonServerConnectionManager too. Lastly,
deprecate the constructors and replace all usages of them in favor of
the Builder

#907