Set the DuplicateCommandHandlerResolver on the SimpleCommandBus
Add the DuplicateCommandHandlerResolver on the SimpleCommandBus by
expanding the Builder (and it's javadoc) accordingly. Also actually use
the resolver if the subscriptions map already contain the given key.
The AsynchronousCommandBus should also be updated, as it inherits from
the SimpleCommandBus. Lastly, a test should be provided which check
whether the resolver is taken in to account

#944