Reduce load on TokenStore from Tracking Event Processor
The Tracking Event Processor will now update the Token before it starts
processing the first Event from a Batch. This will force the necessary
lock on the Token row (when supported) and extends the claim. Only when
event processing took longer than the tokenClaimInterval, the Processor
will explicitly extend the Token's claim. Otherwise, it is assumed the
next batch will be sufficiently quick to start that the token is
extended well before the claim expires.

This behavior can be overridden in the
TrackingEventProcessorConfiguration.

Addresses issue #1168