Add a polling time to the EventBuffer
Add a polling time to the EventBuffer used during the hasNextAvailable
to ensure nobody accidentally performs a hasNextAvailable which will
block for an extended period of time. The nextAvailable operation for
example calls hasNextAvailable with MAX_INTEGER, which initially blocked
 they entire EventBuffer. The `pollingTimeMillis` defaults to 500ms.
Make some clean up adjustments too, like indentation and javadoc.

#event-buffer-improvement