Added ability to add meta data to outgoing queries
The query parameter passed to the QueryGateway was always used as the payload, even if the query already implemented message. This made it impossible to add custom (query-specific) meta data to a query.
This commit checks if the given query implements Message, and uses the payload and meta data if that is the case.