Fixed timing of the invocation of the message monitor on the Simple Event Bus
The SimpleEventBus used a Stream to create a MessageMonitorCallback for each published message. Due to the lazy initialization of streams, this caused the messages to only be marked as "ingested" by the time they we actually dispatched. For monitors performing timing, it may seem that the publication of messages is done in near-0 time.

This commit fixes that, ensuring messages are marked as ingested as soon as they arrive on the Event Bus, and marked as completed (or failed) when the Unit of Work is committed.