Added identifier to TokenStore
This identifier allows the processors to identify whether two TokenStore instances (either on the same node or different nodes) are pointing at the same location. This is useful for AxonServer when trying to balance the number of active threads on nodes. This identifier can distinguish different processors when they carry the same name, but point at different token stores.