Implement registerHandlerEnhancerDefinition
Introduce the registerHandlerEnhancerDefinition implementation. This
slightly adjust the registerHandlerDefinition functionality too, in that
 it appends an additional HandlerDefinition instead of replacing it.
Similar behaviour is followed for the registerHandlerEnhancerDefinition.
Added, a getter should be used for the HandlerDefinition (which pairs
both HandlerDefinitions and HandlerEnhancerDefinitions) to ensure
they've been set as intended prior to usage. Lastly, some warnings can
and should be removed from this class

#1049