Implement registerResourceInjector
Implement a registerResourceInjector. This should introduce a custom
ResourceInjector which is called along side the fixture's default
injector. Both should be called in this scenario to keep ensuring the
correct injection of default components like the CommandBus and EventBus
The custom version can in turn, if implemented as such, override the
default's injected resources

#894