Add assert-no-scheduled-event methods
To provide functionality to match/test for no scheduled events
functions should be introduced which assert that. An assert method
taking in a Duration and a Matcher and an assert method taking in an
Instant and a Matcher should be sufficient. Whilst in the file, resolve
inconsistencies and warnings as mentioned by the IDE

#691