Introduce builder for the EventTypeUpcaster
The EventTypeUpcaster would benefit from builder, with methods called
"from" and "to". Programmatically creating an EventTypeUpcaster would
then make it very clear from which event type to which other event type
this upcaster upcasts. Additionally, introduce a
BuilderUtils#assertNonEmpty, used to validate whether the given
payloadtype is non null and not empty

#1310