Add the LifecycleHandler, it's registration and invocation
Adjust the Configuration API to introduce an onStart/onShutdown which
takes in the LifecycleHandler functional interface. The collections of
start and shutdown handlers should become TreeMaps taking in the phase
as the ordering parameter. A failure during start should result in
throwing a LifecycleHandlerInvocationException and initiations of the
shutdown process. During shutdown the failure should be logged.
Lastly, init handlers no longer need to be phased, and start/shutdown
handlers added out of order should be given precedence.

#891