Add tests, adjust tests and fix some ParameterResolver logic
All the API changes should not have any impact on the existing tests.
However, Spring's wiring logic combined with the
SpringBeanParameterResolverFactory caused for issues. Additionally,
tests should be introduced to cover the changed lifecycle logic in the
DefaultConfigurer, as well as the removal of init-ordering tests in the
DefaultConfigurerTest class

#891