Make default "storeTokenAfterProcessing" value depend on TransactionManager
The previous default of storing tokens before processing (which is a performance improvement), is only valid when event handling is executed in a transaction. Therefore, when configuring the NoTransactionManager, the default behavior should be to claim tokens at the start, and update them after processing a batch. For other TransactionManager implementations, the default remains to store tokens before processing.

Resolves issue #1220