Adjust when AggregateConstructorCommandHandler is used
Currently, we use a AggregateConstructorCommandHandler for constructor
command handlers and regular command handlers annotated with ALWAYS. In
the latter case this wont work, as through this approach the framework
expects an existing instance on those handlers, whilst we clearly don't
have one. An if-else, moving the switch-case in the else and dropping
the use of a AggregateConstructorCommandHandler on the ALWAYS case
will resolve the problem at hand.

#1356