Added support for annotated handler interceptors
This commit adds support for interceptors declared on the handler
instance itself, using the same rules for parameter injection as other
annotated handlers. Depending on the annotation used, handlers are also
given the opportunity to only act on the result of handler execution,
such as a specific type of exception.

Resolves #901