Adjust how the query name will be retrieved
The commit which introduce the option to provide a Message as the query
parameter missed the adjustment to ensure the queryName would stay
consistently. Hence, the queryName could become "Message", which is
definitely not intended. As a resolution, a static method can be
introduced in the GenericMessage class which correctly deduced the
message name. This static method can in turn be used in the QueryGateway
 default methods to supply the right queryName field. Subsequently, the
 tests should be updated to verify this behaviour.

#query-name-fix