Ensure no duplicate events are visible when reading staged events
Events are staged in a Unit of Work when they are published. However, when a Unit of Work is nested, committing a UoW, will stage the events in a parent UoW. When reading staged events during a commit, it is possible that staged events have already been promoted to the parent UoW and are then being read, causing duplicates.

This commit prevents the occurrence of duplicates when reading staged events from a Unit of Work.

Resolved issue #1408