Add activateDefaultTyping(PolymorphicTypeValidator) to JacksonSerializer
In the case users return lists of objects in a Query Handler in
conjunction with using the JacksonSerializer, they will hit the
problem of deserialization missing the exact types to deserialize too.
The JacksonSerializer can benefit to allow easier toggling of the
addition of type information upon serialization to resolve this issue.

#1418