Prevent connections from being shut down if they can be recovered.
This prevents situations where calls fail after a connection was closed by AxonServer because the JVM was unresponsive for a certain period of time. This could be caused by a stop-the-world GC and has also been observed in the case of JVM based Function (e.g. Lambdas) where a JVM is paused to prevent it from consuming resources when inactive.