Introduce EventTrackerStatusChangeListener
Introduce the EventTrackerStatusChangeListener which users will be able
to specify to react to changes in the EventTrackerStatus. They'll
receive a Map of segmentId to EventTrackerStatus if a change occurred.
Additionally, on the interface they can define if they want to react on
updated tracker positions. By default it wont, as those happen often and
 thus would typically not be valuable.

#1338