Adjust reaction to hasNextAvailable slightly
The current solution introduces a form of side effect _why_ it no longer
 loops in the do-while-loop. Namely, subsequent invocations of the
 hasNextAvailable method currently introduce the scenario that the next
 entry from the Iterator is retrieved. This means, the following entry
 is removed from the delegate EventStream. Although this make sure an
 early return occurs, it still does an additional unnecessary loop.
Instead, we should consistently use one way to validate an event is
present, and omit the side effect inducing operation. As this means
we'll no longer perform direct invocations of the EventStream which
could lead to an exception, we should validate whether the stream is
closed prior to proceeding.

#1481