Add exception message if selecting a processing group fails
Although the chances are slim as users would have to provide a lot of
specifics around type selecting, throwing a IllegalStateException
without a message is not helpful when selecting a Class' or Object's
Processing Group fails. Thus make sure a sensible message is shared to
tell the user what's actually failing.

#1500