Add null check before unboxing lastKnownSequence
Although this would typically not occur, it is safer to validate the
lastKnownSequence to be non-null prior to unboxing it to verify it's
size. Sample scenario could be state-stored aggregates where the event
store is missing for some reason. In this case, users might want to
adjust the behaviour of the GenericJpaRepository to set a sequence
number on their own, as they would not be publishing DomainEventMessages
 if the sequence number is missing. Henceforth, they would likely call
 the AnnotatedAggregate#lastSequence method to validate this, returning
 a NullPointerException if this wasn't set every before.

Resolve #1506