Fine tune testing of validating caughtUp flag
The original tests used sleeps to wait for a specified amount of time.
To mitigate this usage, methods which can assert for a given period of
time (with intervals) should be added to the AssertUtils. Further more,
these should be used at the right trigger moment, thus when the event
availability time out kicks in.

#1505