Added root cause exception to stacktrace on error reading events
The EventBuffer would throw an exception when unable to read events from the underlying stream, but it didn't report the original exception if the stream was closed due to an error. This made debugging very hard.
This commit adds the root cause of the closed connection to the exception thrown.