Adjust model factory to be based on inspected types
The meta model factory currently has two problems: (1) it only takes
into account actual classes which contain handler methods and (2) for
each of these it will recursively find all the handlers. Point one
signals that some parts of the Aggregate hierarchy can be completed
disregarded when it comes to searching for aggregate members for
example, as it is perfectly fine for a class to have no handlers and
some members. Point 2 is a hard requirement because _if_ several levels
of the hierarchy contain handlers, then we would recursively find all
fields and methods which should partake in model creation. This however
means that objects on the root will be taken into account every time a
lower level implementation is scanned. We should thus adjust the
behavior to base itself on the inspected classes, and only take the
fields/methods on each of those into account.

#duplicate-child-entities