Ensure usage of gRPC aware Serializer
When retrieving an Aggregate stream, there is a window of opportunity
that the default snapshot/event serializer is used, instead of the
GrpcMetaDataAwareSerializer version. As a result, MetaData might be
missed out on. This should be resolved by overriding the
getSnapshotSerializer/getEventSerializer methods in the
AxonServerEventStore and by using these methods in the
AbstractEventStorageEngine consistently. This issue only comes to light
if snapshotFilters are used, since only then will AxonServer be
requested for snapshots and events separately. The latter caused a call
which doesn't use this specific serializer

#1611