Fixed issue using load or create in combination with Caching repository
In a load-or-create situation, the created aggregate may not have its identifier set, yet. The CachingEventSourcingRepository would fail when attempting to add a created aggregate to the cache, if null keys are not allowed.

This commit removes the addition of entries to the cache when creating an aggregate. Instead, it relies on the entry being added when the aggregate is saved (i.e. when the unit of work is preparing for commit).