Add debug logging when subscribing a command handler
Add debug logging when subscribing a command handler to any CommandBus
implementation. This could clarify if an application isn't subscribing
any command handlers for a users, granted he/she switches on debug
logging. Added, adjust the logging on the distributed CommandBus
implementation to specify a similar log statement is to be expected on
the local segment.

#1246