Separately prepare ChildEntity Command Handlers
In conjunction with the deduplication of ChildEntities within an
Aggregate hierarchy, a ChildEntity's command handlers were no longer
attached to each level within the aggregate hierarchy but only on the
exact class. Although this is correct from a model's perspective, this
is not how the AnnotatedHandlerInspector provides the set of command
handlers for any given type. Such a set includes all handlers on the
concrete type, as well is the levels it implements; thus similarly as
how originally the ChildEntities were added to an AggregateModel.
Although the deduplication of entities is correct, deduplication of the
command/message handlers is simply unfeasible given the number of
locations which expect the full set for a given class type. As such, a
ChildEntity's handlers are no longer added upon finding the entity, but
only after that in the prepareHandlers() method. Here, an entity its
command handlers are added on the given type, but also to any
implementations of that type which belong to the aggregate hierarchy.

#uninvoked-command-handler-on-aggregate-member