Use single QueryHandler implementation on handler registration
Whenever a query handler is registered we should provide the same
QueryHandler instance to the QueryChannel. If we don't, there is a
window of opportunity that query handlers with the same query name but
differing query responses allow for double invocation of the query. This
 occurs because Axon Server does not know of the response type (and
 rightfully so) but only takes into account the query name. If thus two
 query handlers for "findAll" are registered, respectively responding
 with String and Integer, they would now introduce two distinct
 QueryHandler instances which are only stored based on the query name.
Towards AxonServer a subscription is shared which does note the
name/type combination. Thus upon handling of a "findAll" scatter gather
query, the connector receives two messages from AxonServer. The
connector then picks all the QueryHandler instances, which is two, and
invokes both of them for each received message. Making sure the
QueryHandler instance is the same resolves this, as no duplicate
handlers are provided in doing so.

#1663