Improved termination heuristic when response is < batchsize/2 and the resulting event stream has contigious sequence numbers.
I'm unsure if its a bug or by design, but the current sql support can/will fail if a given aggregate has jumps in sequence numbers that are > batchSize.
This feature requires that batchsize be doubled from its current value to retain existing (somewhat questionable) behaviour.

I will admit that I find the batchSize vs maxGapOffset somewhat confusing (the docs on maxGapOffset are beyond my comprehension).
I somehow perceive batchSize to be at a lower layer controlling interactions with a single store. As such
this termination heuristic is applied at the same level.

It might be that someone needs to rethink the overall architecture somewhat around this.