Use result of Condition to decide whether to consume a following event
Instead of invoking nextNode() regardless, we should use the result of
Condition to decide whether to consume a following event. Currently, we
simply wait for the condition to be met, and try to advance regardless.
This means we can technically try to retrieve events twice in quick
succession whilst there's nothing reading for the picking.

#1689

(cherry picked from commit 16b71529472ddb7345bd247ee5dd930dc6bd2206)