Implement reset tokens
Implement reset tokens. First assert resets are supported and that the
processor is no longer running. When that's the case, fetch all the
segments and claim all the tokens subsequently in a transaction. After
this the reset should be initiate, followed up by a token update to
change them to ReplayTokens

#pooled-tracking-event-processor