Implement split segment
Implement the splitSegment operation on the PooledTrackingEventProcessor
The PTEP first validates whether the tokenStore allows splits. If so,
the Coordinator will perform the split. If there's a WorkPackage
corresponding to the segmentId, it will be aborted and it's token split.
If there is no WorkPackage processing the given segmentId, the
Coordinator will try to claim it and split after that. To correctly
perform the split of a trackingToken for a WorkPackage, the
lastStoredToken should be exposed. Furthermore, a static operation on
the TrackerStatus actually performing the split will proof beneficial
to no have to duplicate this process.

#pooled-tracking-event-processor