Propagate errors on cancelled Subscription Query
When a Subscription Query is cancelled the updates Flux is completed
normally instead of with an error. This causes the client to not be able
to recognize whether a Subscription Query was closed due to an error,
or on explicit request by emitting component.

This commit ensures that any errors reported on closing the stream are
propagated into the updates Flux.