Add WorkPackage#stopPackage method
Simply retrieving the lastStoredToken might be an unsafe operation. Even
if it's covered for during it's invocation, it signals danger. Instead
having a dedicated method which stops the WorkPackage and returns the
token of the last event it actually processed is safer. Use this method
during split and merge to base correctly base the split/merge operation

#pooled-tracking-event-processor