Change CoordinatorTask approach
- Return the given CompletableFuture on the run() invocation, to be able
 to attach operations to run on completion
- Make sure the task() methods are protected
- Remove stopPackage() method in favor of simply fetching the token
after aborting a work package. As this branch is only performed _if_
this application holds the claim (the WorkPackage proves this), we can
simply invoke fetchToken in that scenario too.
- Neither SplitTask nor MergeTask so silently return false in case of an
 UnableToClaimTokenException. Instead, the result should complete
 exceptionally in these cases
- Add successful merging log line to the MergeTask
- The Coordinator should give precedence over performing instructions.
To achieve this, attach a reschedule operation on the return
CompletableFuture from CoordinatorTask#run. After this, return
immediately. This unblocks the Coordinator thread, but also ensure the
tasks are performed first.
- Rename startCoordinationTask() to scheduleImmediateCoordinationTask()
- Fix tests according to provided changes

#pooled-tracking-event-processor