Add test case for event availability callback
The EventStream might support an event availability callback. That way,
the Coordinator is automatically nudged to immediately start processing
events again. In absence of this, it will simply reschedule itself at a
later stage. Using the callback enhance processing efficiency (if the
stream supports this functionality). The implementation in the
Coordinator to this end should be tested

#pooled-tracking-event-processor