Use sensitive defaults in ProcessingInstructionHelper
Instead of reverting to the default value of 0 for numeric processing instructions, the helper should return values that are sensible for each individual type of instruction.
In particular the "number of results" instruction should default to 1 when not provided.

(cherry picked from commit 282a460d4a333fa6ee433af04044eb15d8e32d99)