Introduce a jakarta specific validator
Segregate the BeanValidationInterceptor from any jakarta specifics, by
porting it back to javax. We cannot simply adjust this old component to
jakarta, as that constitutes a breaking change. Instead a jakarta
specific implementation (and exception) should be provided. This can be
used by users which want to move away from javax.

#javax-jakarta-validation