Use isStreamingProcessor flag
Update the axonserver-connector version to 4.5-SNAPSHOT, so that we can
work with the updated EventProcessorInfo object. On this we can define
whether the processor for which information is shared is a streaming
instance yes/no. This flag dictates whether the consumer of this
information will show Tracking Status and the Streaming operations (e.g.
 split and merge)

#1712