Throw AxonConfigurationException if aggregate member type = an interface
The AnnotatedAggregateModel cannot guarantee that all the command
handlers of an interface-type aggregate member are scanned. This holds,
as it is uncertain which concrete type is actually present as the
aggregate member, thus its message handlers would be ignored. The
behaviour was to (faultily) throw a `NullPointerException` whenever the
command handlers were retrieved from an Aggregate Member. Instead we
should throw an AxonConfigurationException with a clear message stating
a concrete type should be used in the aggregate member annotation.

#aggregate-member-interface-type