Add aggregate type
As stated in the SnapshotFilter documentation, the aggregate type is a
mandatory validation field. This is required as all SnapshotFilters
of each aggregate type are combined. Thus a simple revision check will
impact every snapshot search. As such, if the type does not match, the
filter MUST return true.

fix/revision-snapshot-filter