Add a boolean to validate if the WorkerLauncher is still running
The adjusted start/shutdown mechanics, including a count of the active
processor threads, have brought ot light that sometimes threads are
segments are still claimed for processing even though the TEP-shutdown
operation has completed successfully. The only process constructing
these threads, is the private WorkerLauncher class. By adding a
validator noting when the WorkerLauncher is active and adding that next
to invocations of activeProcessorThreads(), we can overcome the issue of
accidentally stating the shutdown is finalized even though the
WorkerLauncher is still active.

#workerlauncher-thread-inclusion