Adjust upcaster ordering logic
The AnnotationAwareOrderComparator does not check for the annotation on
bean creation methods, but only on the classes themselves. We can
instead use the findAnnotationOnBean(beanName, Order.class) method to
get the ordering from any bean creation spot. This allows us to support
the @Order annotation on @Bean annotated methods or fields too.

#1892