Adjust secured type search approach
Using the ComponentScan annotation is, sadly, not recommended by the
Spring team. Instead, basing ourselves around the EntityScan annotation
is more reasonable. Added, in absence of the EntityScan annotation,
using the AutoConfigurationPackages#get(BeanFactory) method is regarded
as the default in absence of that annotation. A short description on
why, together with the new approach, can be found here ->
https://stackoverflow.com/questions/50808941/how-to-get-basepackages-of
-componentscan-programatically-at-runtime

#1963