Adjust approach to updating tokens in absence of usable events
Adjust the if-available-events check to a while-available-events. Add a
validation to the processing deadline, that's the processing start time
plus the eventAvailabilityTimeout. Doing so makes this deadline act
similar to the extension-claim-threshold of the PooledStreaming
processor. Lastly, attach a check whether events were added to the
batch, as otherwise the loop will continue adding events instead of
eagerly handle newfound events.

#token-update-improvement