Ensure cached aggregates are invalidated on failures
The cache was not invalidated in all cases, which made it possible that subsequent commands would hit a wrong instance of an aggregate. Typically, the failure of these commands would in turn invalidate the entry.

This commit ensures that all points where cached entries could be in use properly register an onRollback handler to invalidate the cached entry.