Add sequenceIdentifier method to the EventHandlerInvoker
A dead lettering EventHandlerInvoker should be able to retrieve the
sequence id for a given event from its delegate. Without this, it will
not be able to correctly store the events in the right order. By added
EventHandlerInvoker#sequenceIdentifier, any specific implementation can
retrieve the seqId for any event from its delegate.

#2021