Add constructor with parameter specific message
Add a constructor that ingests a QueryMessage and provides an exception
for which the message is specific on what's being matched on. Doing so,
this solution mirrors the approach for the NoHandlerForCommandException.
Thus, the new constructor should be used whenever the
MessageHandlerMember#canHandle method is invoked for a query handler.

#1646