Replace coverage-aggregate profile for separate coverage project
As is described in the shared links below, the jacoco:report-aggregate
goal requires to be run after all projects carrying coverage data have
been build. To achieve this, we require a separate pom.xml to generate
the aggregate coverage report. To that end, the coverage-report/pom.xml
is introduced. This replaces the coverage-aggregate profile.
Additionally, the coverage profile is expanded to enable the
coverage-report/pom.xml too. Lastly, both the coverage and
integration-test profile can be activated with the coverage property.

Here are a couple of links that led to the solution that's introduced:
- Maven Multi-Module discussion doc: https://github.com/jacoco/jacoco/wiki/MavenMultiModule
- jacoco:report-aggregate doc: https://www.eclemma.org/jacoco/trunk/doc/report-aggregate-mojo.html
- MultiModule usage forum thread: https://groups.google.com/g/jacoco/c/FpdLbxsXSTY
- jacoco:report-aggregate sample: https://github.com/jacoco/jacoco/tree/master/jacoco-maven-plugin.test/it/it-report-aggregate

#feature/sonar-report-aggregation