Solidify API of the DeadLetterQueue
- Introduce the concept of a QueueIdentifier (plus concrete
implementations), pairing the identifier and group fields used in the
DeadLetterQueue.
- Let enqueue and enqueueIfPresent return a DeadLetterEntry and
Optional<DeadLetterEntry. This allows for immediate validation if
operations succeeded, as well as additional referral to the entry
constructed.
- Separate maxSize() into maxQueues() and maxQueueSize(). This allows
space for a maximum amount of queues as well as a maximum queue size.
This requires the isFull method to take in a QueueIdentifier, to be able
 to make the correct validation.
- Peek should contain the group it is peeking for. Otherwise, results
from different (processing) groups may end up in the wrong handlers.
Next to that, let peek return an Optional to eliminate null values.
- Remove DeadLetterQueue#evaluationSucceeded, in favor of
DeadLetterEntry#release.
- Remove DeadLetterQueue#evaluationFailed. Instead, the implementations
should update the DeadLetterEntry#expiresAt and reinsert the entry,
right after performing peek. This safeguards against concurrent
retrieval of entries.
- Introduce a means to clear out the DeadLetterQueue.
- Make all necessary adjustments to committed code (e.g., the
DeadLetteringEventHandlerInvoker and InMemoryDeadLetterQueue) to
comply with the above description.

#2021