Let DeadLetteringEventHandlerInvoker extend SimpleEventHandlerInvoker
The DeadLetteringEventHandlerInvoker should extend the
SimpleEventHandlerInvoker for a couple of reasons. One, we don't have a
clean way to disregard event handling, and thus event enqueueing, for
events that aren't intended for this segment. Thus, multithreaded
solutions might all invoke the DeadLetterQueue#enqueueIfPresent whilst
they will not handle the event to begin with. Secondly, it removes the
necessity to make the sequenceIdentifier public logic of the
EventHandlerInvoker. Thirdly, it allows far greater flexibility to
adjust the overall handle approach of the SimpleEventHandlerInvoker,
ensuring the correct flow of matching segments, enqueueing if present,
handling the event, enqueuing on failures and (potentially) dealing with
 the ErrorHandler too.

#2021