Redesign of Spring Boot Auto Configuration support
Redesigned the Spring Boot Auto Configuration mechanism to fully use "regular" bean definitions and post processors, rather than ImportBeanDefinitionRegistrars that also check for the availability of beans.

A subclass of the DefaultConfigurer, the SpringConfigurer, is used to locate possible components in the Spring Application context, instead of hard-configuring any possible type of component.

The classes using the existing approach have been deprecated.