Introduce IntermediateEventRepresentation#getContentType
Although the provided fix works, it means we need to expand the if/else
branch for any new serializer approach. Instead of the branch, it would
be nice if the IntermediateEventRepresentation exposes the content type
of the data object it's holding. That way, the EventTypeUpcaster can
depend on the previous type, not adjusting it for further upcasters in
the chain.

#bug/event-type-upcaster